backend/requirements/prod.txt
# =============================================================================
# PRODUCTION DEPENDENCIES - Optimized for production deployment
# =============================================================================

# Include base requirements
-r base.txt

# Production Server
gunicorn==21.2.0

# Production Monitoring & APM
sentry-sdk[fastapi]==1.40.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Security Headers & Middleware
secure==0.3.0

# Production Database Drivers
asyncpg==0.29.0  # PostgreSQL async driver
redis==5.0.1     # Redis for caching/sessions

# Health Checks
py-healthcheck==1.10.1

# Production Logging
python-json-logger==2.0.7
structlog==24.1.0

# AWS Integration (if using AWS)
# boto3==1.34.25

# Google Cloud Integration (if using GCP)
# google-cloud-logging==3.9.0
# google-cloud-monitoring==2.18.1
